feat: add email authentication option
