+++ src/components/AuthVisual.tsx
@@ -1,168 +1,28 @@
 import React, { useState } from 'react';
-import { View, Text, TextInput, TouchableOpacity, StyleSheet, Alert, useColorScheme } from 'react-native';
-import auth from '@react-native-firebase/auth';
+import { View, Text, TouchableOpacity, StyleSheet, useColorScheme } from 'react-native';
 import GoogleLogin from './GoogleAuth.tsx';
+import EmailAuth from './EmailAuth';
 import { getColors, Colors } from '../theme/colors';
 
 const AuthScreen = () => {
-  const [isRegistering, setIsRegistering] = useState(false);
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [isResettingPassword, setIsResettingPassword] = useState(false);
-  const [resetEmail, setResetEmail] = useState('');
+  const [showEmailAuth, setShowEmailAuth] = useState(false);
   const isDarkMode = useColorScheme() === 'dark';
   const colors = getColors(isDarkMode);
   const styles = createStyles(colors);
-  const handleAuth = async () => {
-    try {
-      console.log('🔧 handleAuth spustený. isRegistering:', isRegistering);
 
-      if (isRegistering) {
-        const response = await fetch('http://10.0.2.2:3001/api/register', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({ email, password }),
-        });
-
-        const result = await response.json();
-
-        if (!response.ok) {
-          console.error('❌ Backend error:', result);
-          Alert.alert('Chyba registrácie', result?.error || 'Neznáma chyba');
-          return;
-        }
-
-        Alert.alert(
-          'Registrácia úspešná',
-          'Na tvoju emailovú adresu bol odoslaný overovací email. Po jeho potvrdení sa môžeš prihlásiť.',
-        );
-
-        // ŽIADNE signOut ani nič iné – používateľ nebol nikdy prihlásený
-
-      } else {
-        // 🔑 PRIHLÁSENIE
-        const userCredential = await auth().signInWithEmailAndPassword(email, password);
-        const user = userCredential.user;
-
-        if (!user.emailVerified) {
-          Alert.alert(
-            'Email nie je overený',
-            'Prosím, over svoju emailovú adresu pred prihlásením.',
-          );
-          await auth().signOut();
-          return;
-        }
-
-        const idToken = await user.getIdToken();
-
-        await fetch('http://10.0.2.2:3001/api/auth', {
-          method: 'POST',
-          headers: {
-            Authorization: `Bearer ${idToken}`,
-            'Content-Type': 'application/json',
-          },
-        });
-
-        Alert.alert('Úspech', 'Prihlásenie úspešné');
-      }
-    } catch (error: any) {
-      console.error('❌ Globálna chyba:', error);
-      Alert.alert('Chyba', error.message || 'Neznáma chyba');
-    }
-  };
-  if (isResettingPassword) {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.title}>Obnoviť heslo</Text>
-
-        <TextInput
-          placeholder="Zadaj svoj email"
-          value={resetEmail}
-          onChangeText={setResetEmail}
-          style={styles.input}
-          keyboardType="email-address"
-          autoCapitalize="none"
-        />
-
-        <TouchableOpacity
-          style={styles.button}
-          onPress={async () => {
-            try {
-              const response = await fetch('http://10.0.2.2:3001/api/reset-password', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ email: resetEmail }),
-              });
-
-              const result = await response.json();
-
-              if (!response.ok) {
-                Alert.alert('Chyba', result?.error || 'Nepodarilo sa odoslať email');
-                return;
-              }
-
-              Alert.alert('Hotovo', 'Email na obnovu hesla bol odoslaný');
-              setIsResettingPassword(false);
-              setResetEmail('');
-            } catch (error) {
-              console.error('❌ Chyba pri resete hesla:', error);
-              Alert.alert('Chyba', 'Zlyhanie pri odoslaní emailu');
-            }
-          }}
-        >
-          <Text style={styles.buttonText}>Odoslať link</Text>
-        </TouchableOpacity>
-
-        <TouchableOpacity onPress={() => setIsResettingPassword(false)}>
-          <Text style={styles.link}>← Späť na prihlásenie</Text>
-        </TouchableOpacity>
-      </View>
-    );
+  if (showEmailAuth) {
+    return <EmailAuth onBack={() => setShowEmailAuth(false)} />;
   }
+
   return (
     <View style={styles.container}>
-    <Text style={styles.title}>{isRegistering ? 'Registrácia' : 'Prihlásenie'}</Text>
-
-      <TextInput
-  placeholder="Email"
-  value={email}
-  onChangeText={setEmail}
-  style={styles.input}
-  keyboardType="email-address"
-  autoCapitalize="none"
-  />
-
-  <TextInput
-    placeholder="Heslo"
-  value={password}
-  onChangeText={setPassword}
-  secureTextEntry
-  style={styles.input}
-  />
-
-  <TouchableOpacity onPress={handleAuth} style={styles.button}>
-  <Text style={styles.buttonText}>
-    {isRegistering ? 'Registrovať sa' : 'Prihlásiť sa'}
-    </Text>
-    </TouchableOpacity>
-
-    <TouchableOpacity onPress={() => setIsRegistering(!isRegistering)}>
-  <Text style={styles.link}>
-  {isRegistering
-    ? 'Už máš účet? Prihlás sa'
-    : 'Nemáš účet? Zaregistruj sa'}
-  </Text>
-  </TouchableOpacity>
-      <View style={styles.socialLoginContainer}>
-        <GoogleLogin />
-      </View>
-      <TouchableOpacity onPress={() => setIsResettingPassword(true)}>
-        <Text style={styles.link}>Zabudol si heslo?</Text>
+      <Text style={styles.title}>Prihlásenie</Text>
+      <GoogleLogin />
+      <TouchableOpacity style={styles.button} onPress={() => setShowEmailAuth(true)}>
+        <Text style={styles.buttonText}>Použiť email</Text>
       </TouchableOpacity>
-  </View>
-);
+    </View>
+  );
 };
 
 const createStyles = (colors: Colors) =>
@@ -172,12 +32,6 @@ const createStyles = (colors: Colors) =>
       justifyContent: 'center',
       padding: 24,
       backgroundColor: colors.background,
-    },
-    socialLoginContainer: {
-      marginTop: 20,
-      paddingVertical: 10,
-      borderTopWidth: 1,
-      borderTopColor: colors.border,
       alignItems: 'center',
     },
     title: {
@@ -187,31 +41,19 @@ const createStyles = (colors: Colors) =>
       textAlign: 'center',
       color: colors.text,
     },
-    input: {
-      borderWidth: 1,
-      borderColor: colors.border,
-      padding: 12,
-      borderRadius: 8,
-      marginBottom: 16,
-      backgroundColor: colors.cardBackground,
-      color: colors.text,
-    },
     button: {
       backgroundColor: colors.primary,
       padding: 14,
       borderRadius: 8,
-      marginBottom: 12,
+      marginTop: 20,
+      width: '100%',
     },
     buttonText: {
       color: '#fff',
       textAlign: 'center',
       fontWeight: '600',
     },
-    link: {
-      color: colors.primary,
-      textAlign: 'center',
-      marginTop: 12,
-    },
   });
 
 export default AuthScreen;
+
